Moved rate limiter config to the middleware/ folder
